refactor(sync): own historical staging service
