+++ knowledge/features/sync/overview.md
@@ -21,7 +21,7 @@ sources:
     title: SyncTuning constants
     last_modified: 2026-08-02
   - id: historical-sync
-    resource: ../../../lib/database/maintenance.dart
+    resource: ../../../lib/features/sync/services/historical_sync_service.dart
     title: Historical sync staging and failed-row retry
     last_modified: 2026-08-05
   - id: current-architecture
@@ -289,7 +289,8 @@ contract is in
 The agent phases stamp before they send. An agent entity or link persisted with
 `vectorClock: null` is applied by the receiving peer but skipped by
 `_recordReceivedAgentEntity`, so it lands invisible to the sequence log, gap
-detection and backfill. `reSyncInterval`'s agent `enqueueAction` therefore
+detection and backfill. `HistoricalSyncService.reSyncInterval`'s agent
+`enqueueAction` therefore
 gives any clockless row a vector clock and persists it *before* enqueueing —
 inside the interval sweep, so a *Last 30 days* run repairs only what it is
 about to send rather than every legacy row in the database. The stamp is
